requirements.txt
python-telegram-bot>=21.6
python-dotenv>=1.0.0
httpx>=0.27.0
openai>=1.50.0
anthropic>=0.40.0
google-genai>=1.5.0
